refactor: modularize RichTextEditor by extracting toolbar, constants, and utility functions into dedicated files.
