+++ components/common/RichTextEditor/constants.ts
@@ -1,0 +1,9 @@
+export const LIST_TYPES = ['numbered-list', 'bulleted-list'] as const;
+export const TEXT_ALIGN_TYPES = ['left', 'center', 'right', 'justify'] as const;
+
+export const HOTKEYS = {
+  'mod+b': 'bold',
+  'mod+i': 'italic',
+  'mod+u': 'underline',
+  'mod+`': 'code',
+};
